gymnast/requirements.txt
asgiref==3.2.3
decorator==4.4.1
Django==2.2.8
djangorestframework==3.11.0
networkx==2.4
numpy==1.17.4
pytz==2019.3
redis==3.3.11
sqlparse==0.3.0
